requirements.txt
contourpy==1.2.1
cycler==0.12.1
einops==0.8.0
filelock==3.14.0
fonttools==4.51.0
fsspec==2024.3.1
ftfy==6.2.0
fvcore==0.1.5.post20221221
imageio==2.34.1
iopath==0.1.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==24.0
Pillow==9.5.0
portalocker==2.8.2
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.5.10
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
simplejson==3.19.2
six==1.16.0
ssd==1.0.0
sympy==1.12
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.10
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.11.0
wcwidth==0.2.13
yacs==0.1.8
